Added 705A Solution

+++ Codeforces/1030A.cpp
@@ -6,13 +6,13 @@ int main()
     int n, flag = 0;
     cin >> n;
     int arr[n];
-    for (size_t i = 0; i < n; i++)
+    for (int i = 0; i < n; i++)
     {
         /* code */
         cin >> arr[i];
     }
 
-    for (size_t i = 0; i < n; i++)
+    for (int i = 0; i < n; i++)
     {
         /* code */
         if (arr[i] == 1)
